+++ src/devices/transitions/animate.ts
