+++ src/devices/transitions/easings.ts
