+++ src/types/src/__tests__/osc.spec.ts
@@ -1,5 +1,5 @@
 import { OSCEasingType } from '..'
-import { Easing } from '../../../easings'
+import { Easing } from '../../../devices/transitions/easings'
 
 describe('OSC', () => {
 	test('OSC Easing type', () => {
